requirements.txt
anyio==3.6.1
certifi==2022.6.15
colorzero==2.0
gpiozero==1.6.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
notion-client==1.0.0
python-dotenv==0.20.0
pytz==2022.1
rfc3986==1.5.0
RPi.GPIO==0.7.1
sniffio==1.2.0
